Seleccion

+++ src/code/commandCenter.c
@@ -122,10 +122,11 @@ void createRequest (char *namepipe, book bookRequest, int fd){
 
 int main (int argc, char *argv[]){
 
-
-   if (argc != 5){
+   printf("%i", argc);
+   if (argc != 5 && argc != 3){
       perror("\tNumero de argumentos invalidos\n");
-      printf("\tej: ./debug/center –i files/PS.txt –p debug/pipeReceptor \n");
+      printf("\tej: ./debug/center [–i files/PS.txt] –p debug/pipeReceptor \n");
+      printf("\t[] argumento opcional \n");
       exit (0);
    }
    
@@ -136,37 +137,28 @@ int main (int argc, char *argv[]){
    mode_t fifo_mode = S_IRUSR | S_IWUSR;
    sprintf(namePS, "debug/PS#%i", getpid());
 
-   //comprobar si argc es 2 se entran los valores por consola pero si es 4 se lee el archivo y no se pregunta el menu
-   printf("\n");
-   printf("\t___Welcome to BookTime___\n");
-   printf("\t1. Leer el archivo\n");
-   printf("\t2. Digitar la informacion del libro\n");
-   printf("\tEscoja una opcion: ");
-   scanf("\t%d", &option);
+   if(argc == 5){
 
-   printf("\tSe crea el pipe %s para recibir la respuesta del receptor\n", namePS);
-   printf("\t---------------------\n");
-   
-   unlink(namePS); 
-   if (mkfifo (namePS, fifo_mode) == -1) {
-      perror("\t[!]Request mkfifo");
-      exit(1);
-   }
+      printf("\tSe crea el pipe %s para recibir la respuesta del receptor\n", namePS);
+      printf("\t---------------------\n");
+      
+      unlink(namePS); 
+      if (mkfifo (namePS, fifo_mode) == -1) {
+         perror("\t[!]Request mkfifo");
+         exit(1);
+      }
 
-   printf("\tSe abre el %s para enviar proceso a realizar\n", argv[4]);
-   printf("\t---------------------\n");
+      printf("\tSe abre el %s para enviar proceso a realizar\n", argv[4]);
+      printf("\t---------------------\n");
 
-   do { 
-      if ((fd = open(argv[4], O_WRONLY))== -1) {
-         perror("\t[!] Error en la lectura del pipe\n");
-         printf("\tSe volvera a intentar despues\n");
-	      sleep(5);        
-      } else create = 1;
-   } while (create == 0);
+      do { 
+         if ((fd = open(argv[4], O_WRONLY))== -1) {
+            perror("\t[!] Error en la lectura del pipe\n");
+            printf("\tSe volvera a intentar despues\n");
+            sleep(5);        
+         } else create = 1;
+      } while (create == 0);
 
-   switch (option){
-
-   case 1:
       readFile(argv[2]);
       printf("\tse leyo del archivo %s\n", argv[2]);
       for(int i = 0; i < pos; i++){
@@ -182,81 +174,96 @@ int main (int argc, char *argv[]){
          printf("\tPipe: %s\n", booksRequests[i].secondpipe);
          createRequest(argv[4], booksRequests[i], fd);
       }
-      break;
-         case 2: 
-            do{
-               printf("\n");
-               printf("\t_______BookTime______\n");
-               printf("\n");
-               printf("\t1. Pedir Libro\n");
-               printf("\t2. Devolver Libro\n");
-               printf("\t3. Renovar Libro\n");
-               printf("\t0. Salir\n");
-               printf("\tEscoja una opcion: ");
-               scanf("\t%i", &optionMenu);
+   } else if(argc == 3){
 
-               switch (optionMenu){
-               case 1:
-                  printf("\n");
-                  printf("\tInserte el nombre del libro\n");
-                  fflush(stdin);
-                  scanf("%s",&bookRequest.name);
-                  printf("\tInserte el ISBN del libro\n");
-                  scanf("%d", &bookRequest.ISBN);
-                  bookRequest.operation = 'P';
-                  strcpy(bookRequest.secondpipe, namePS);
-                  printf("\t---------------------\n");
-                  printf("\tSe crea solicitud\n");
-                  printf("\tSolicitud: %c\n", bookRequest.operation);
-                  printf("\tNombre: %s\n", bookRequest.name);
-                  printf("\tISBN: %d\n", bookRequest.ISBN);
-                  printf("\tPipe: %s\n", bookRequest.secondpipe);
-                  createRequest(argv[4], bookRequest, fd);
-                  break;
-               case 2:
-                  printf("\n");
-                  printf("\tInserte el nombre del libro\n");
-                  fflush(stdin);
-                  scanf("%s",&bookRequest.name);
-                  printf("\tInserte el ISBN del libro\n");
-                  scanf("%d", &bookRequest.ISBN);
-                  bookRequest.operation = 'D';
-                  strcpy(bookRequest.secondpipe, namePS);
-                  printf("\t---------------------\n");
-                  printf("\tSe crea solicitud\n");
-                  printf("\tSolicitud: %c\n", bookRequest.operation);
-                  printf("\tNombre: %s\n", bookRequest.name);
-                  printf("\tISBN: %d\n", bookRequest.ISBN);
-                  printf("\tPipe: %s\n", bookRequest.secondpipe);
-                  createRequest(argv[4], bookRequest, fd);
-                  break;
-               case 3:
-                  printf("\n");
-                  printf("\tInserte el nombre del libro\n");
-                  fflush(stdin);
-                  scanf("%s",&bookRequest.name);
-                  printf("\tInserte el ISBN del libro\n");
-                  scanf("%d", &bookRequest.ISBN);
-                  bookRequest.operation = 'R';
-                  strcpy(bookRequest.secondpipe, namePS);
-                  printf("\t---------------------\n");
-                  printf("\tSe crea solicitud\n");
-                  printf("\tSolicitud: %c\n", bookRequest.operation);
-                  printf("\tNombre: %s\n", bookRequest.name);
-                  printf("\tISBN: %d\n", bookRequest.ISBN);
-                  printf("\tPipe: %s\n", bookRequest.secondpipe);
-                  createRequest(argv[4], bookRequest, fd);
-                  break;
-               default:
-                  printf("\n");
-                  printf("\tOpcion no valida\n");
-                  break;
-               }
-            }while(optionMenu != 0);
-         break;
-   default:
-      break;
+      printf("\tSe crea el pipe %s para recibir la respuesta del receptor\n", namePS);
+      printf("\t---------------------\n");
+         
+      unlink(namePS); 
+      if (mkfifo (namePS, fifo_mode) == -1) {
+         perror("\t[!]Request mkfifo");
+         exit(1);
+      }
+      printf("\tSe abre el %s para enviar proceso a realizar\n", argv[2]);
+      printf("\t---------------------\n");
+      do { 
+         if ((fd = open(argv[2], O_WRONLY))== -1) {
+            perror("\t[!] Error en la lectura del pipe\n");
+            printf("\tSe volvera a intentar despues\n");
+            sleep(5);        
+         } else create = 1;
+      } while (create == 0);
+      do{
+         printf("\n");
+         printf("\t_______BookTime______\n");
+         printf("\n");
+         printf("\t1. Pedir Libro\n");
+         printf("\t2. Devolver Libro\n");
+         printf("\t3. Renovar Libro\n");
+         printf("\t0. Salir\n");
+         printf("\tEscoja una opcion: ");
+         scanf("\t%i", &optionMenu);
+
+         switch (optionMenu){
+         case 1:
+            printf("\n");
+            printf("\tInserte el nombre del libro\n");
+            fflush(stdin);
+            scanf("%s",&bookRequest.name);
+            printf("\tInserte el ISBN del libro\n");
+            scanf("%d", &bookRequest.ISBN);
+            bookRequest.operation = 'P';
+            strcpy(bookRequest.secondpipe, namePS);
+            printf("\t---------------------\n");
+            printf("\tSe crea solicitud\n");
+            printf("\tSolicitud: %c\n", bookRequest.operation);
+            printf("\tNombre: %s\n", bookRequest.name);
+            printf("\tISBN: %d\n", bookRequest.ISBN);
+            printf("\tPipe: %s\n", bookRequest.secondpipe);
+            createRequest(argv[4], bookRequest, fd);
+            break;
+            case 2:
+            printf("\n");
+            printf("\tInserte el nombre del libro\n");
+            fflush(stdin);
+            scanf("%s",&bookRequest.name);
+            printf("\tInserte el ISBN del libro\n");
+            scanf("%d", &bookRequest.ISBN);
+            bookRequest.operation = 'D';
+            strcpy(bookRequest.secondpipe, namePS);
+            printf("\t---------------------\n");
+            printf("\tSe crea solicitud\n");
+            printf("\tSolicitud: %c\n", bookRequest.operation);
+            printf("\tNombre: %s\n", bookRequest.name);
+            printf("\tISBN: %d\n", bookRequest.ISBN);
+            printf("\tPipe: %s\n", bookRequest.secondpipe);
+            createRequest(argv[4], bookRequest, fd);
+            break;
+         case 3:
+            printf("\n");
+            printf("\tInserte el nombre del libro\n");
+            fflush(stdin);
+            scanf("%s",&bookRequest.name);
+            printf("\tInserte el ISBN del libro\n");
+            scanf("%d", &bookRequest.ISBN);
+            bookRequest.operation = 'R';
+            strcpy(bookRequest.secondpipe, namePS);
+            printf("\t---------------------\n");
+            printf("\tSe crea solicitud\n");
+            printf("\tSolicitud: %c\n", bookRequest.operation);
+            printf("\tNombre: %s\n", bookRequest.name);
+            printf("\tISBN: %d\n", bookRequest.ISBN);
+            printf("\tPipe: %s\n", bookRequest.secondpipe);
+            createRequest(argv[4], bookRequest, fd);
+            break;
+         default:
+            printf("\n");
+            printf("\tOpcion no valida\n");
+            break;
+         }
+      }while(optionMenu != 0);
    }
+
    close(fd);
    printf("\tEl proceso %s termina\n", namePS);
    exit(0);
